day/requirements.txt
django
pillow
openpyxl
djangorestframework
